app/requirements/production.txt

-r requirements.txt
gunicorn
django-allauth
